requirements.txt
aiogram==3.13.1
python-dotenv==1.1.0
aiosqlite==0.20.0